Fix device_info name
