+++ custom_components/skyq/entity.py
@@ -1,20 +1,25 @@
 """Entity representing a Sky Device."""
 
-from homeassistant.helpers.entity import Entity
-
 from .const import DOMAIN
 
 
-class SkyQEntity(Entity):
+class SkyQEntity:
     """Representation of a SkyQ Device."""
 
+    def __init__(self, remote, config):
+        """Initialise the SkyQEntity."""
+        self._remote = remote
+        self._config = config
+        self._unique_id = config.unique_id
+        self._deviceInfo = None
+
     @property
-    def device_info(self):
+    def skyq_device_info(self):
         """Entity device information."""
         return (
             {
                 "identifiers": {(DOMAIN, self._deviceInfo.serialNumber)},
-                "name": self.name,
+                "name": self._config.name,
                 "manufacturer": self._deviceInfo.manufacturer,
                 "model": self._deviceInfo.hardwareModel,
                 "sw_version": f"{self._deviceInfo.ASVersion}:{self._deviceInfo.versionNumber}",
@@ -22,3 +27,14 @@ class SkyQEntity(Entity):
             if self._deviceInfo
             else None
         )
+
+    async def _async_get_device_info(self, hass):
+        """Query the device for device info."""
+        if self._deviceInfo:
+            return
+        self._deviceInfo = await hass.async_add_executor_job(self._remote.getDeviceInformation)
+        if self._deviceInfo:
+            if not self._unique_id:
+                self._unique_id = self._deviceInfo.epgCountryCode + "".join(
+                    e for e in self._deviceInfo.serialNumber.casefold() if e.isalnum()
+                )
